feat: fetch config from s3

+++ scripts/build-new-version.mjs
@@ -1,19 +1,31 @@
+import dotenv from 'dotenv';
 import fs from 'fs';
 import { execSync } from 'child_process';
-const files = fs.readdirSync('./configs/services-schema');
-const schemas = {};
 
 run();
 
 function run() {
+  dotenv.config({ path: './.env' });
+  const AWS_S3_BUCKET = process.env.AWS_S3_BUCKET;
+  const AWS_PROFILE = process.env.AWS_PROFILE;
+
+  if (!AWS_S3_BUCKET || !AWS_PROFILE) {
+    throw new Error('AWS_S3_BUCKET and AWS_PROFILE must be set');
+  }
+  execSync(
+    `aws s3 sync ${AWS_S3_BUCKET}/configs/service-schema/ ./configs/service-schema/ --profile ${AWS_PROFILE}`,
+  );
   execSync(`
     mkdir -p src/ts-types
     rm -rf src/ts-types/*.*
   `);
-  generateTsTypes();
+
+  const files = fs.readdirSync('./configs/service-schema');
+
+  generateTsTypes(files);
 }
 
-function generateTsTypes() {
+function generateTsTypes(files) {
   const tab = '  ';
   let index = '';
   for (const file of files) {
@@ -66,12 +78,7 @@ function toCamelCase(string) {
 }
 
 function _load(file) {
-  if (schemas[file]) {
-    return schemas[file];
-  }
-  const schema = JSON.parse(
-    fs.readFileSync(`./configs/services-schema/${file}`, 'utf8'),
+  return JSON.parse(
+    fs.readFileSync(`./configs/service-schema/${file}`, 'utf8'),
   );
-  schemas[file] = schema;
-  return schema;
 }
